libepoxy: propagate buildInputs
Some packages that depend on libepoxy fail to build due to libGL and
libX11 not being present during the build. One example is `qemu_xen`
(or really any QEMU built with GL support), as it's built with libepoxy,
which links against libGL, but since libGL isn't present, the build fails.

Signed-off-by: Fernando Rodrigues <alpha@sigmasquadron.net>